requirements.txt
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.1
google-genai>=1.0.0
line-bot-sdk>=3.5.0
notion-client>=2.2.0
python-dotenv>=1.0.0
